enable line breaks at hyphens in \url{...}
This is an option of the url package. That package is loaded by hyperref if not already loaded, which makes loading it after either a no-op, or if options are given, an error. So it has to be moved to before loading hyperref. In general, all packages that hyperref fiddles with should be loaded before hyperref, according to the hyperref documentation.